pyproject.toml
[tool.poetry]
name = "primed"
version = "0.1.0"
description = ""
authors = ["Sky \"g3ner1c\" Hong <g3ner1c-sky@pm.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
beautifulsoup4 = "^4.12.2"
aiohttp = "^3.9.1"
motor = "^3.3.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
isort = "^5.13.2"
black = "^23.12.0"
tox = "^4.11.4"
python-dotenv = "^1.0.0"
